scansavvy-backend/requirements.txt
fastapi
uvicorn
motor
bcrypt
python-dotenv
pydantic
